Don't set local ip as ELASTICSEARCH_HOST
This version of broker deploys only to tsuru, hence it doesn't need functionality to find out and set local IP as elasticserver_host. This was needed when you deploy the broker directly on the elasticsearch server. We have then also renamed ELASTICSEACH_IP to ELASTICSEARCH_HOST to reflect the changes.