Update healthcheck
Simplify the healthcheck. We were not parsing the output of _cluster/health?pretty=true, hence no need to obtain it. Just do a simple check to / to see if elasticsearch is responding at all. Return 204 if yes, otherwise 500. The intention of _cluster/health?pretty=true was to pass the json returned by elasticsearch to user, so that they can determine themselves the health, but tsuru doesn't pass the body. Thus we return empty body.